Don't propagate CommsEffect in core_map
