+++ jax/_src/pallas/primitives.py
@@ -1136,16 +1136,6 @@ def semaphore_signal(
   )
 
 
-# TODO(sharadmv,ivyzheng): remove this once we use axis dicts primarily
-class CommsEffect(effects.Effect):
-  pass
-_comms_effect = CommsEffect()
-effects.lowerable_effects.add_type(CommsEffect)
-effects.control_flow_allowed_effects.add_type(CommsEffect)
-effects.remat_allowed_effects.add_type(CommsEffect)
-effects.custom_derivatives_allowed_effects.add_type(CommsEffect)
-
-
 @semaphore_signal_p.def_effectful_abstract_eval
 def _semaphore_signal_abstract_eval(
     *avals,
@@ -1169,9 +1159,8 @@ def _semaphore_signal_abstract_eval(
     for aval in device_id_flat_avals:
       if aval.dtype != jnp.dtype("int32"):
         raise ValueError("`device_id`s must be an int32 value.")
-    effs.add(_comms_effect)
+    effs.add(pallas_core.comms_effect)
   return [], effs
-
 
 def _semaphore_signal_pp_eqn(eqn: jax_core.JaxprEqn,
                              context: jax_core.JaxprPpContext,
